Add timeout blocks for Terraform resources (#97)
* add timeouts

By convention (https://developer.hashicorp.com/terraform/language/resources/syntax#operation-timeouts)
terraform resources can have timeouts associated with their operations
through `timeout` blocks.
This change sets the default timeout to 2 minutes, except where other
defaults were hard-coded before and allows users to specify timeout
blocks without causing parsing errors.

Use the contexts to setup and propagate deadlines for each operation
according to what is set in the terraform file. Since each API call
happens within the same context, the same deadline applies, giving each
operation (create, update, read, delete) as a whole the configured
timeout rather than setting this timeout on individual API calls of the
operation.

Where applicable, the timeout of resources is measured against the
observed changes of the API resources or events rather than completion
of API calls. This is particularly important for resources where
operations can take a lot longer than the API call.
E.g. for the Delete operation on a Volume, the time until a matching
event is observed is counted. For the Update operation of a
VirtualMachine, the time (if necessary) until the VM resource has been
restarted and is observed in running state is counted.
For a contrast, the Delete operation on a storage class happens almost
immediately and the timeout is only counted against the API call.

fixes: https://github.com/harvester/harvester/issues/5632

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

* cloud-init secret: no wait on delete

Don't wait for a state change when deleting a cloud-init secret. The
cloud-init secret is just a normal core/v1 Secret API object. These
don't have a state, they either exist or they don't but there is not any
intermediate state like the is with a VM or a Pod. As a consequence,
it's not necessary to wait for a state change when deleting a secret.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

---------

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>