Fix Volume Resource
Fix delete procedure for volume resource. Upon deletion, the volume
(PVC) does not produce a usable event. Thus watching the K8s resource
for state change will never yeild the expected result. As a consequence
the Terraform provider will eventually time-out and produce an error.
The fix is to watch for a resource state change instead of an event.
This will propagate the deletion of the resource appropriately and the
Terraform provider will succeed with the deletion.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>