Add Volume Delete Test Step
Add test step ensuring that volumes are deleted correctly.
This test works independent of the timeout since in case the timeout is
exceeded the delete operation will already be considered a fail.
Thus, if the test step succeeds, the delete would not have timed out.
The test step also ensures that the resource is actually removed.

related-to: harvester/harvester#6415

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>