Terraform Provider: Kubeconfig as Base64 string
Accept kubeconfig as base64 string.
When configuring the Harvester Terraform provider, the kubeconfig can
either be given as path to a file, from which to read, or the contents
can be supplied directly as base64 string.
This allows for greater flexibility in supplying this sensitive data to
the Terraform provider.

related-to: harvester/harvester#6234

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>