Add support for setting the boot order on network interfaces (#122)
* bump Longhorn

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

* bump Harvester

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

* bump Harvester

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

* VM Network Interface: Expose Boot Order Property

Expose the `BootOrder` property of network interfaces of a virtual
machine. This is necessary to allow the creation of VMs which can boot
from network/with PXE.

related-to: https://github.com/harvester/harvester/issues/7695

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

* work around Harvester release

v1.5.0 did not contain the fixes to the VMBuilder interface that I had
tried to put in there. As a result, the Terraform provider must be built
off of a ref on the master branch.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

* fix linter errors

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

* fix tests

Golang tests have a default timeout of 10 minutes for the whole test
suite. This isn't enough for acceptance testing as those tests require a
few minutes each.
Therefore, when acceptance testing, the timeout needs to be disabled.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

* add test for network interface boot order

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

* fix style issues

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

* fix schema

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

---------

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>