Add IPPool and LoadBalancer resources (#102)
* Add IPPool and LoadBalancer resources

related-to: harvester/harvester#4813
related-to: harvester/harvester#4814

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

* Enable tests

Enable tests, add test scaffolding for IPPool resource, speed up
compilation

Signed-Off-By: Moritz Röhrich <moritz.roehrich@suse.com>
Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

* IPPool and LoadBalancer

More functionality, especially around the LoadBalancer resource

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

* Execute tests

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

* Fix Volume Resource

Fix delete procedure for volume resource. Upon deletion, the volume
(PVC) does not produce a usable event. Thus watching the K8s resource
for state change will never yeild the expected result. As a consequence
the Terraform provider will eventually time-out and produce an error.
The fix is to watch for a resource state change instead of an event.
This will propagate the deletion of the resource appropriately and the
Terraform provider will succeed with the deletion.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

* Fix tests

Fix VM Network and Storage Class tests

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

* Add negative tests for IPPool and Loadbalancer

Add negative test cases for IPPool and LoadBalancer resources. This
ensures that if mandatory parameters are not given, the program exits
with an appropriate error.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

* Add VM tags test, LoadBalancer test

Add acceptance tests for VirtualMachine labels (tags) and LoadBalancer
resources. The labeling/tagging mechnism is crucial for these tests to
work, since a LoadBalancer is required by the admission webhook to have
at least one VM that matches its selectors.

Also improve documentation.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

* fix tests

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

* fix type cast

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

* fix formatting for examples

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

* fix docs

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>

---------

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>
Signed-off-by: Moritz Röhrich <moritz.roehrich@suse.com>