add checksum field to image resource schema (#133)
Add checksum field to image resource schema and constructor. The
checksum allows users to make sure an image from a download source is
downloaded correctly and the content are as expected.

Changing the checksum field forces the image CRD to be re-created,
because only this forces the image to be downloaded again and the
checksum to be re-calculated and compared with the updated value.

related-to: harvester/harvester#8265

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>