Feature: Custom Labels on any Resource
Add a block `labels` to all Terraform resources. This block allows
setting almost arbitrary labels on the Kubernetes CRs corresponding to
each Terraform resource.
The only restriction on these labels is that a) they are valid
Kubernetes labels and b) that the keys don't have either the
tags.harvesterhci.io/ nor the harvesterhci.io prefixes.

The `labels` block is implemented as a map, meaning that the syntax for
adding labels to the resources is as follows, e.g.:

```
resource "harvester_image" "leap" {
  [...]

  labels = {
    key = "value"
    foo = "bar"
  }

  [...]
}
```

The ability to manage labels in a flexible manner is paramount to
several other features of Harvester. For example, the backend selector
of the load-balancer relies heavily on the use of labels to identify
which virtual machines to route traffic to.

related-to: harvester/harvester#4816

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>